Move canonical identity RuntimeWorld methods into canonical_world.rs
Pure, behavior-preserving extraction from main.rs toward ENG.md priority #1
(tracked by #255): the nine canonical identity/version RuntimeWorld methods
(ensure_canonical_identities, canonical_ref, runtime_handle_for_canonical_ref,
canonical_pact_ref, entity_version, event_entity_refs,
bump_entity_versions_for_events, refresh_canonical_events,
refresh_all_canonical_events) move beside the CanonicalIdentityState types
they operate on, as a pub(crate) split-impl. main.rs shrinks 63611 -> 63391
and the line ceiling lowers to lock in the win. Tests remain in main.rs and
canonical_identity_tests.rs for this slice.
